backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.0.3
supabase==2.3.4
pydantic[email]==2.5.3
PyJWT==2.8.0 